# PERF: Avoid using `@recursive`.