@charris 2021-08-03
Hmm, that is *all* the uses of `@recursive`, might as well remove the definition also. @mattip Thoughts?