@anntzer 2021-08-03
I don't have a preference wrt. completely deleting the implementation of `@recursive`, just let me know what you prefer either way.