@seberg 2021-08-03
Please delete it, its just unnecessary garbage code now, it is not like it has any immediate use (or probably even any use at all).